Add reflect parsed filter to component logic
