+++ src/components/search/formFields/ProductFilter.vue
@@ -1,6 +1,7 @@
 <template>
   <div v-if="data.length > 0">
     <check-buttons
+      ref="checkButtons"
       v-for="(item, idx) in data"
       :key="idx"
       :data="item"
@@ -11,7 +12,14 @@
 </template>
 
 <script lang="ts">
-import { Vue, Component, Prop, PropSync, Watch } from "vue-property-decorator";
+import {
+  Vue,
+  Component,
+  Prop,
+  PropSync,
+  Watch,
+  Ref,
+} from "vue-property-decorator";
 import CheckButtons from "../../common/inputs/CheckButtons.vue";
 import { Filter, RequestFilter } from "../../../types/model";
 import { objectMap } from "../../../utils";
@@ -25,13 +33,14 @@ export default class ProductFilter extends Vue {
   private readonly data!: Filter[];
   @PropSync("filter", { required: true })
   private filterSync!: RequestFilter;
+  @Ref()
+  private readonly checkButtons!: CheckButtons[];
 
   private checkedFilter: InstanceFilter = {};
 
   @Watch("data")
   changedData(value: Filter[]): void {
     if (value.length > 0) this.initFilter();
-    this.$forceUpdate();
   }
 
   @Watch("filterSync")
@@ -39,6 +48,9 @@ export default class ProductFilter extends Vue {
     this.checkedFilter = objectMap<string, string[]>(val, (value) =>
       value.split(",")
     );
+    this.$nextTick(() => {
+      this.checkButtons.forEach((component) => component.initDefaultValue());
+    });
   }
 
   private initFilter() {
